requirements.txt
# Core RAG dependencies
llamaindex==0.10.11
pinecone-client==2.2.4
openai==1.6.1

# Database
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3

# PDF processing
pypdf==3.17.4

# Web framework
streamlit==1.29.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1